AI Doctor/requirements-clean.txt
# Core FastAPI and server dependencies
fastapi==0.115.6
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI and ML dependencies
groq==0.15.0
elevenlabs==1.50.3

# Audio processing
gtts==2.5.4
speech-recognition==3.11.0
pydub==0.25.1
pyaudio==0.2.14

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.32.3

# Image processing
Pillow==10.2.0

# Optional: For enhanced audio processing
ffmpeg-python==0.2.0
